client/tests/python_selenium/requirements.txt
selenium==4.24.0
webdriver-manager==4.0.1
pytest==8.3.3
pytest-html==4.1.1
python-dotenv==1.0.1
requests==2.32.3


